chore: remove journal aggregate, Hasura proxy, and old codegen
- Remove entire journal API and views (ui/src/api/journal/, ui/src/views/journal/)
- Remove Hasura v1 proxy from server (WithApiV1Proxy, hasura_backend flag/env)
- Remove old Hasura-based codegen (codegen.ts → replaced with gqlgen config,
  codegen.next.ts deleted, gql/gql.ts, gql/graphql.ts, gql/index.ts deleted)
- Rename documents.next.ts → documents.ts for incident, layer, message APIs
- Remove VITE_API_WS_URL from .env (no more WebSocket / Hasura subscription)
- Clean up Journal type, journals field on Incident, IncidentContext journal state,
  and all test fixtures referencing journals
