--- .github/workflows/hasura-workflow.yml
@@ -1,77 +1,0 @@
-name: Hasura Build & Push
-permissions:
-  contents: read
-
-on:
-  workflow_call:
-    inputs:
-      push:
-        description: 'Whether to push the hasura image'
-        required: false
-        default: false
-        type: boolean
-
-jobs:
-  hasura:
-    runs-on: ubuntu-24.04-arm
-    name: "Build and Push Hasura Image"
-    permissions:
-      actions: write
-      contents: read
-      packages: write
-      id-token: write
-    steps:
-      - name: Checkout
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-
-      - name: Set up QEMU
-        uses: docker/setup-qemu-action@96fe6ef7f33517b61c61be40b68a1882f3264fb8 # v4.2.0
-      
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@37fe631027851001ddb9b187196cc803df7f5f0e # v4.3.0
-        with:
-          # Enable driver-specific options for improved caching
-          driver-opts: |
-            image=moby/buildkit:latest
-            network=host
-      
-      - name: Docker meta for hasura
-        id: meta-hasura
-        uses: docker/metadata-action@dc802804100637a589fabce1cb79ff13a1411302 # v6.2.0
-        with:
-          images: ghcr.io/f-eld-ch/sitrep-hasura
-          tags: |
-            type=edge,branch=develop
-            type=raw,value=stable,enable=${{ github.ref == format('refs/heads/{0}', 'main') }}
-            type=ref,event=branch
-            type=ref,event=pr,prefix=pr-,enable=true,priority=600
-            type=semver,pattern={{version}},prefix=v,priority=420
-            type=semver,pattern={{major}}.{{minor}},prefix=v,priority=410
-            type=semver,pattern={{major}},prefix=v,priority=400
-            type=sha
-      
-      - name: Login to GHCR
-        if: inputs.push
-        uses: docker/login-action@dbcb813823bdd20940b903addbd779551569679f # v4.6.0
-        with:
-          registry: ghcr.io
-          username: ${{ github.actor }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-      
-      - name: Build and push hasura image
-        uses: docker/build-push-action@53b7df96c91f9c12dcc8a07bcb9ccacbed38856a # v7.3.0
-        with:
-          context: .
-          file: Dockerfile.hasura
-          platforms: linux/amd64,linux/arm64
-          push: ${{ inputs.push }}
-          tags: ${{ steps.meta-hasura.outputs.tags }}
-          labels: ${{ steps.meta-hasura.outputs.labels }}
-          # Improved caching configuration with branch-specific scope
-          cache-from: |
-            type=gha,scope=hasura-${{ github.ref_name }}
-            type=gha,scope=hasura-main
-          cache-to: type=gha,mode=max,scope=hasura-${{ github.ref_name }}
-          # Add provenance and SBOM attestation
-          provenance: mode=max
-          sbom: true
