+++ ui/.env
@@ -1,4 +1,3 @@
 VITE_API_URL="/api/v2/graphql"
-VITE_API_WS_URL="ws://localhost:4180/v1/graphql"
 VITE_VERSION=${VERSION}
 VITE_SHA_VERSION=${GIT_SHA}
