+++ ui/codegen.ts
@@ -1,29 +1,32 @@
 /**
- * graphql-codegen config for the CURRENT Hasura schema.
+ * graphql-codegen config for the FUTURE gqlgen schema.
  *
- * Schema source: hasura/schema/hasura.graphql — a committed SDL snapshot
- * generated by introspecting the running Hasura engine as role "editor":
+ * Pointed at api/schema.graphql — the shared SDL contract.
+ * Run: yarn codegen:next
  *
- *   yarn codegen:schema
+ * Purpose: validate that GraphQL documents written for the future gqlgen schema
+ * are compatible with api/schema.graphql WHILE Hasura is still serving
+ * production. Any document the Go server cannot answer fails at codegen time,
+ * not at runtime after cutover.
  *
- * Introspect as "editor" (not admin) so codegen only validates documents
- * that can actually be executed at runtime.
+ * Documents convention: name them *.next.ts alongside the current *.documents.ts.
+ * The current Hasura documents use Hasura-specific vocabulary (byPk, _eq, uuid,
+ * affectedRows, …) that does not exist in the future schema and must never be
+ * included here. As each aggregate is ported (Phases 3–6), its future documents
+ * are added as a *.next.ts file and the corresponding error count in this check
+ * drops to zero — a measurable burn-down.
  *
- * Committing the generated output (src/gql/) keeps CI hermetic: there is
- * no Hasura dependency in the build pipeline. The generated diff on a PR is
- * the wire-contract change, reviewable by engineers.
- *
- * Run: yarn codegen
- * Check (CI): yarn codegen:check
+ * Output goes to src/gql/next/ — not imported by the app yet, just validated.
+ * When cutover is complete, this config becomes the primary codegen.ts.
  */
 
 import type { CodegenConfig } from "@graphql-codegen/cli";
 
 const config: CodegenConfig = {
-  schema: "../hasura/schema/hasura.graphql",
+  schema: "../api/schema.graphql",
   documents: ["src/api/**/*.ts"],
   generates: {
-    "./src/gql/": {
+    "./src/gql/next/": {
       preset: "client",
       presetConfig: {
         fragmentMasking: false,
@@ -32,16 +35,9 @@ const config: CodegenConfig = {
         useTypeImports: true,
         strictScalars: true,
         scalars: {
-          uuid: "string",
-          timestamptz: {
-            input: "Date | string",
-            output: "string",
-          },
-          jsonb: "Record<string, unknown>",
-          point: "string",
-          MediumEnum: "string",
-          PriorityStatusEnum: "string",
-          TriageStatusEnum: "string",
+          DateTime: "string",
+          Geometry: "GeoJSON.Geometry | null",
+          JSONObject: "Record<string, unknown>",
         },
         avoidOptionals: {
           field: true,
