+++ ui/src/App.tsx
@@ -1,5 +1,5 @@
 import { lazy, Suspense, useEffect } from "react";
-import { Navigate, RouterProvider, createBrowserRouter, useParams } from "react-router";
+import { Navigate, RouterProvider, createBrowserRouter } from "react-router";
 
 import "./App.scss";
 
@@ -9,7 +9,7 @@ import { Spinner } from "components";
 import { useTranslation } from "react-i18next";
 import { IncidentContextProvider, UserProvider } from "utils";
 import { Editor as IncidentEditor, List as IncidentList, New as IncidentNew } from "views/incident";
-import { Editor as JournalEditor, List as JournalMessageList } from "views/journal";
+import { Editor as JournalEditor, List as JournalMessageList, New as JournalNew, Overview as JournalOverview } from "views/journal";
 import { Layout, LayoutMarginLess } from "views/Layout";
 import { List as ImmediateMeasuresList } from "views/measures/immediateMeasures";
 import { List as RequestList } from "views/measures/requests";
@@ -25,12 +25,6 @@ import it from "dayjs/locale/it";
 import LocalizedFormat from "dayjs/plugin/localizedFormat";
 
 const MapView = lazy(() => import("views/map"));
-
-// Redirects old /journal/:journalId[/…] URLs — journalId is dropped, incidentId preserved.
-function JournalIdRedirect({ to }: { to: string }) {
-  const { incidentId } = useParams<{ incidentId: string }>();
-  return <Navigate to={`/incident/${incidentId}/journal/${to}`} replace />;
-}
 
 const router = createBrowserRouter([
   {
@@ -62,33 +56,6 @@ const router = createBrowserRouter([
                 <IncidentEditor />
               </Layout>
             ),
-          },
-          {
-            path: "journal",
-            children: [
-              {
-                path: "edit",
-                element: (
-                  <Layout>
-                    <JournalEditor />
-                  </Layout>
-                ),
-              },
-              {
-                path: "messages",
-                element: (
-                  <Layout>
-                    <JournalMessageList showControls={false} autoScroll={true} />
-                  </Layout>
-                ),
-              },
-              // Redirects for legacy routes
-              { path: "view", element: <Navigate to="messages" replace /> },
-              { path: "new", element: <Navigate to="edit" replace /> },
-              { path: ":journalId", element: <JournalIdRedirect to="messages" /> },
-              { path: ":journalId/edit", element: <JournalIdRedirect to="edit" /> },
-              { path: ":journalId/messages", element: <JournalIdRedirect to="messages" /> },
-            ],
           },
           {
             path: "resources",
@@ -131,6 +98,29 @@ const router = createBrowserRouter([
                 <ImmediateMeasuresList />
               </Layout>
             ),
+          },
+          {
+            path: "journal",
+            children: [
+              { index: true, element: <JournalOverview /> },
+              {
+                path: "edit",
+                element: (
+                  <Layout>
+                    <JournalEditor />
+                  </Layout>
+                ),
+              },
+              {
+                path: "messages",
+                element: (
+                  <Layout>
+                    <JournalMessageList showControls={false} autoScroll={true} />
+                  </Layout>
+                ),
+              },
+              { path: "new", element: <JournalNew /> },
+            ],
           },
         ],
       },
