--- ui/src/api/journal/index.ts
@@ -1,5 +1,0 @@
-export type { CreateJournalArgs } from "./commands";
-export { useCloseJournal, useCreateJournal, useReopenJournal } from "./commands";
-export { afterJournalWrite } from "./invalidate";
-export type { JournalsData } from "./queries";
-export { useJournals } from "./queries";
