--- ui/src/api/journal/invalidate.ts
@@ -1,5 +1,0 @@
-import { GET_JOURNALS } from "./documents";
-
-export function afterJournalWrite(incidentId: string) {
-  return [{ query: GET_JOURNALS, variables: { incidentId } }];
-}
