--- ui/src/gql/gql.ts
@@ -1,154 +1,0 @@
-/* eslint-disable */
-import * as types from './graphql';
-import type { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/core';
-
-/**
- * Map of all GraphQL operations in the project.
- *
- * This map has several performance disadvantages:
- * 1. It is not tree-shakeable, so it will include all operations in the project.
- * 2. It is not minifiable, so the string of a GraphQL query will be multiple times inside the bundle.
- * 3. It does not support dead code elimination, so it will add unused operations.
- *
- * Therefore it is highly recommended to use the babel or swc plugin for production.
- * Learn more about it here: https://the-guild.dev/graphql/codegen/plugins/presets/preset-client#reducing-bundle-size
- */
-type Documents = {
-    "\n    query FetchIncidents {\n      incidents(orderBy: { createdAt: DESC }, where: { deletedAt: { _isNull: true } }) {\n        id\n        name\n        createdAt\n        updatedAt\n        deletedAt\n        closedAt\n        location {\n          name\n          coordinates\n        }\n      }\n    }\n  ": typeof types.FetchIncidentsDocument,
-    "\n  query GetIncidentDetail($incidentId: uuid!) {\n    incidentsByPk(id: $incidentId) {\n      id\n      name\n      createdAt\n      closedAt\n      updatedAt\n      location {\n        id\n        name\n        coordinates\n      }\n      divisions {\n        id\n        name\n        description\n      }\n      journals {\n        id\n        name\n      }\n    }\n  }\n": typeof types.GetIncidentDetailDocument,
-    "\n  mutation InsertIncident(\n    $name: String!\n    $location: String\n    $divisions: [DivisionsInsertInput!]!\n    $journalName: String\n    $layerName: String\n  ) {\n    insertIncidentsOne(\n      object: {\n        name: $name\n        location: { data: { name: $location } }\n        journals: { data: { name: $journalName } }\n        layers: { data: { name: $layerName } }\n        divisions: { data: $divisions }\n      }\n    ) {\n      id\n      name\n      journals {\n        id\n        name\n      }\n      divisions {\n        name\n        id\n        description\n      }\n      layers {\n        name\n        id\n      }\n    }\n  }\n": typeof types.InsertIncidentDocument,
-    "\n  mutation UpdateIncident(\n    $incidentId: uuid!\n    $name: String!\n    $location: String!\n    $locationId: uuid!\n    $divisions: [DivisionsInsertInput!]!\n  ) {\n    updateLocationsByPk(pkColumns: { id: $locationId }, _set: { name: $location }) {\n      id\n      name\n    }\n    insertDivisions(\n      objects: $divisions\n      onConflict: { constraint: divisions_name_incident_id_key, updateColumns: [description, name] }\n    ) {\n      affectedRows\n    }\n    updateIncidentsByPk(pkColumns: { id: $incidentId }, _set: { name: $name }) {\n      id\n      name\n      journals {\n        id\n        name\n      }\n      divisions {\n        name\n        id\n        description\n      }\n    }\n  }\n": typeof types.UpdateIncidentDocument,
-    "\n  mutation CloseIncident($incidentId: uuid, $closedAt: timestamptz) {\n    updateJournals(\n      where: { incident: { id: { _eq: $incidentId } }, closedAt: { _isNull: true } }\n      _set: { closedAt: $closedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        closedAt\n      }\n    }\n    updateIncidents(where: { id: { _eq: $incidentId } }, _set: { closedAt: $closedAt }) {\n      affectedRows\n      returning {\n        id\n        closedAt\n      }\n    }\n  }\n": typeof types.CloseIncidentDocument,
-    "\n  mutation DeleteIncident($incidentId: uuid, $deletedAt: timestamptz) {\n    updateJournals(\n      where: { incident: { id: { _eq: $incidentId } }, deletedAt: { _isNull: true } }\n      _set: { deletedAt: $deletedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        deletedAt\n      }\n    }\n    updateIncidents(\n      where: {\n        id: { _eq: $incidentId }\n        deletedAt: { _isNull: true }\n        closedAt: { _isNull: false }\n      }\n      _set: { deletedAt: $deletedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        deletedAt\n      }\n    }\n  }\n": typeof types.DeleteIncidentDocument,
-    "\n  query GetJournals($incidentId: uuid) {\n    incidents(where: { id: { _eq: $incidentId } }) {\n      id\n      name\n      journals(orderBy: { createdAt: ASC }) {\n        id\n        name\n        createdAt\n        updatedAt\n        closedAt\n        deletedAt\n      }\n    }\n  }\n": typeof types.GetJournalsDocument,
-    "\n  mutation InsertJournal($name: String!, $incidentId: uuid!) {\n    insertJournalsOne(object: { incidentId: $incidentId, name: $name }) {\n      id\n      name\n      createdAt\n      updatedAt\n      closedAt\n      deletedAt\n    }\n  }\n": typeof types.InsertJournalDocument,
-    "\n    mutation CloseJournal($journalId: uuid, $closedAt: timestamptz) {\n      updateJournals(where: { id: { _eq: $journalId } }, _set: { closedAt: $closedAt }) {\n        affectedRows\n        returning {\n          id\n          closedAt\n        }\n      }\n    }\n  ": typeof types.CloseJournalDocument,
-    "\n  query GetLayers($incidentId: uuid!) {\n    layers(where: { incidentId: { _eq: $incidentId } }) {\n      id\n      name\n      features {\n        id\n        geometry\n        properties\n        createdAt\n        updatedAt\n        deletedAt\n      }\n    }\n  }\n": typeof types.GetLayersDocument,
-    "\n  mutation AddFeature($layerId: uuid!, $id: uuid!, $geometry: jsonb, $properties: jsonb) {\n    insertFeaturesOne(\n      object: { layerId: $layerId, id: $id, geometry: $geometry, properties: $properties }\n    ) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n": typeof types.AddFeatureDocument,
-    "\n  mutation UpdateFeature($id: uuid!, $geometry: jsonb, $properties: jsonb) {\n    updateFeaturesByPk(\n      pkColumns: { id: $id }\n      _set: { geometry: $geometry, properties: $properties }\n    ) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n": typeof types.UpdateFeatureDocument,
-    "\n  mutation DeleteFeature($id: uuid!, $deletedAt: timestamptz) {\n    updateFeaturesByPk(pkColumns: { id: $id }, _set: { deletedAt: $deletedAt }) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n": typeof types.DeleteFeatureDocument,
-    "\n  mutation AddLayer($incidentId: uuid!, $name: String!) {\n    insertLayersOne(object: { incidentId: $incidentId, name: $name }) {\n      id\n    }\n  }\n": typeof types.AddLayerDocument,
-    "\n  query GetMessages($journalId: uuid!) {\n    journalsByPk(id: $journalId) {\n      incident {\n        id\n        divisions {\n          id\n          name\n          description\n        }\n      }\n    }\n    messages(\n      where: { journal: { id: { _eq: $journalId } }, deletedAt: { _isNull: true } }\n      orderBy: { time: DESC }\n    ) {\n      id\n      content\n      sender\n      receiver\n      senderDetail\n      receiverDetail\n      medium: mediumId\n      time\n      createdAt\n      updatedAt\n      deletedAt\n      divisions {\n        division {\n          id\n          name\n          description\n        }\n      }\n      triageId\n      priorityId\n    }\n  }\n": typeof types.GetMessagesDocument,
-    "\n  query GetMessageForTriage($messageId: uuid!) {\n    messagesByPk(id: $messageId) {\n      id\n      content\n      sender\n      receiver\n      senderDetail\n      receiverDetail\n      medium: mediumId\n      time\n      divisions {\n        division {\n          id\n          name\n          description\n        }\n      }\n      createdAt\n      updatedAt\n      deletedAt\n      triageId\n      priorityId\n      journal {\n        incident {\n          divisions {\n            id\n            name\n            description\n          }\n        }\n      }\n    }\n  }\n": typeof types.GetMessageForTriageDocument,
-    "\n    mutation InsertMessage(\n      $journalId: uuid\n      $sender: String\n      $receiver: String\n      $time: timestamptz\n      $content: String\n      $receiverDetail: String\n      $senderDetail: String\n      $medium: MediumEnum\n    ) {\n      insertMessagesOne(\n        object: {\n          content: $content\n          journalId: $journalId\n          receiver: $receiver\n          sender: $sender\n          time: $time\n          mediumId: $medium\n          senderDetail: $senderDetail\n          receiverDetail: $receiverDetail\n        }\n      ) {\n        id\n        createdAt\n        content\n        receiver\n        sender\n        senderDetail\n        receiverDetail\n        medium: mediumId\n        time\n        updatedAt\n        triageId\n        priorityId\n        divisions {\n          division {\n            name\n          }\n        }\n        deletedAt\n      }\n    }\n  ": typeof types.InsertMessageDocument,
-    "\n    mutation UpdateMessage(\n      $messageId: uuid!\n      $content: String\n      $sender: String\n      $receiver: String\n      $time: timestamptz\n      $receiverDetail: String\n      $senderDetail: String\n      $medium: MediumEnum\n    ) {\n      updateMessagesByPk(\n        pkColumns: { id: $messageId }\n        _set: {\n          content: $content\n          sender: $sender\n          receiver: $receiver\n          time: $time\n          mediumId: $medium\n          senderDetail: $senderDetail\n          receiverDetail: $receiverDetail\n        }\n      ) {\n        id\n        createdAt\n        content\n        receiver\n        sender\n        senderDetail\n        receiverDetail\n        medium: mediumId\n        time\n        updatedAt\n        triageId\n        priorityId\n        divisions {\n          division {\n            name\n          }\n        }\n        deletedAt\n      }\n    }\n  ": typeof types.UpdateMessageDocument,
-    "\n  mutation SaveMessageTriage(\n    $messageId: uuid!\n    $priority: PriorityStatusEnum\n    $triage: TriageStatusEnum\n    $messageDivisions: [MessageDivisionInsertInput!]!\n  ) {\n    deleteMessageDivision(where: { messageId: { _eq: $messageId } }) {\n      affectedRows\n    }\n    insertMessageDivision(objects: $messageDivisions) {\n      affectedRows\n    }\n    updateMessagesByPk(\n      pkColumns: { id: $messageId }\n      _set: { priorityId: $priority, triageId: $triage }\n    ) {\n      id\n      divisions {\n        division {\n          name\n        }\n      }\n      triageId\n      priorityId\n    }\n  }\n": typeof types.SaveMessageTriageDocument,
-};
-const documents: Documents = {
-    "\n    query FetchIncidents {\n      incidents(orderBy: { createdAt: DESC }, where: { deletedAt: { _isNull: true } }) {\n        id\n        name\n        createdAt\n        updatedAt\n        deletedAt\n        closedAt\n        location {\n          name\n          coordinates\n        }\n      }\n    }\n  ": types.FetchIncidentsDocument,
-    "\n  query GetIncidentDetail($incidentId: uuid!) {\n    incidentsByPk(id: $incidentId) {\n      id\n      name\n      createdAt\n      closedAt\n      updatedAt\n      location {\n        id\n        name\n        coordinates\n      }\n      divisions {\n        id\n        name\n        description\n      }\n      journals {\n        id\n        name\n      }\n    }\n  }\n": types.GetIncidentDetailDocument,
-    "\n  mutation InsertIncident(\n    $name: String!\n    $location: String\n    $divisions: [DivisionsInsertInput!]!\n    $journalName: String\n    $layerName: String\n  ) {\n    insertIncidentsOne(\n      object: {\n        name: $name\n        location: { data: { name: $location } }\n        journals: { data: { name: $journalName } }\n        layers: { data: { name: $layerName } }\n        divisions: { data: $divisions }\n      }\n    ) {\n      id\n      name\n      journals {\n        id\n        name\n      }\n      divisions {\n        name\n        id\n        description\n      }\n      layers {\n        name\n        id\n      }\n    }\n  }\n": types.InsertIncidentDocument,
-    "\n  mutation UpdateIncident(\n    $incidentId: uuid!\n    $name: String!\n    $location: String!\n    $locationId: uuid!\n    $divisions: [DivisionsInsertInput!]!\n  ) {\n    updateLocationsByPk(pkColumns: { id: $locationId }, _set: { name: $location }) {\n      id\n      name\n    }\n    insertDivisions(\n      objects: $divisions\n      onConflict: { constraint: divisions_name_incident_id_key, updateColumns: [description, name] }\n    ) {\n      affectedRows\n    }\n    updateIncidentsByPk(pkColumns: { id: $incidentId }, _set: { name: $name }) {\n      id\n      name\n      journals {\n        id\n        name\n      }\n      divisions {\n        name\n        id\n        description\n      }\n    }\n  }\n": types.UpdateIncidentDocument,
-    "\n  mutation CloseIncident($incidentId: uuid, $closedAt: timestamptz) {\n    updateJournals(\n      where: { incident: { id: { _eq: $incidentId } }, closedAt: { _isNull: true } }\n      _set: { closedAt: $closedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        closedAt\n      }\n    }\n    updateIncidents(where: { id: { _eq: $incidentId } }, _set: { closedAt: $closedAt }) {\n      affectedRows\n      returning {\n        id\n        closedAt\n      }\n    }\n  }\n": types.CloseIncidentDocument,
-    "\n  mutation DeleteIncident($incidentId: uuid, $deletedAt: timestamptz) {\n    updateJournals(\n      where: { incident: { id: { _eq: $incidentId } }, deletedAt: { _isNull: true } }\n      _set: { deletedAt: $deletedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        deletedAt\n      }\n    }\n    updateIncidents(\n      where: {\n        id: { _eq: $incidentId }\n        deletedAt: { _isNull: true }\n        closedAt: { _isNull: false }\n      }\n      _set: { deletedAt: $deletedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        deletedAt\n      }\n    }\n  }\n": types.DeleteIncidentDocument,
-    "\n  query GetJournals($incidentId: uuid) {\n    incidents(where: { id: { _eq: $incidentId } }) {\n      id\n      name\n      journals(orderBy: { createdAt: ASC }) {\n        id\n        name\n        createdAt\n        updatedAt\n        closedAt\n        deletedAt\n      }\n    }\n  }\n": types.GetJournalsDocument,
-    "\n  mutation InsertJournal($name: String!, $incidentId: uuid!) {\n    insertJournalsOne(object: { incidentId: $incidentId, name: $name }) {\n      id\n      name\n      createdAt\n      updatedAt\n      closedAt\n      deletedAt\n    }\n  }\n": types.InsertJournalDocument,
-    "\n    mutation CloseJournal($journalId: uuid, $closedAt: timestamptz) {\n      updateJournals(where: { id: { _eq: $journalId } }, _set: { closedAt: $closedAt }) {\n        affectedRows\n        returning {\n          id\n          closedAt\n        }\n      }\n    }\n  ": types.CloseJournalDocument,
-    "\n  query GetLayers($incidentId: uuid!) {\n    layers(where: { incidentId: { _eq: $incidentId } }) {\n      id\n      name\n      features {\n        id\n        geometry\n        properties\n        createdAt\n        updatedAt\n        deletedAt\n      }\n    }\n  }\n": types.GetLayersDocument,
-    "\n  mutation AddFeature($layerId: uuid!, $id: uuid!, $geometry: jsonb, $properties: jsonb) {\n    insertFeaturesOne(\n      object: { layerId: $layerId, id: $id, geometry: $geometry, properties: $properties }\n    ) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n": types.AddFeatureDocument,
-    "\n  mutation UpdateFeature($id: uuid!, $geometry: jsonb, $properties: jsonb) {\n    updateFeaturesByPk(\n      pkColumns: { id: $id }\n      _set: { geometry: $geometry, properties: $properties }\n    ) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n": types.UpdateFeatureDocument,
-    "\n  mutation DeleteFeature($id: uuid!, $deletedAt: timestamptz) {\n    updateFeaturesByPk(pkColumns: { id: $id }, _set: { deletedAt: $deletedAt }) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n": types.DeleteFeatureDocument,
-    "\n  mutation AddLayer($incidentId: uuid!, $name: String!) {\n    insertLayersOne(object: { incidentId: $incidentId, name: $name }) {\n      id\n    }\n  }\n": types.AddLayerDocument,
-    "\n  query GetMessages($journalId: uuid!) {\n    journalsByPk(id: $journalId) {\n      incident {\n        id\n        divisions {\n          id\n          name\n          description\n        }\n      }\n    }\n    messages(\n      where: { journal: { id: { _eq: $journalId } }, deletedAt: { _isNull: true } }\n      orderBy: { time: DESC }\n    ) {\n      id\n      content\n      sender\n      receiver\n      senderDetail\n      receiverDetail\n      medium: mediumId\n      time\n      createdAt\n      updatedAt\n      deletedAt\n      divisions {\n        division {\n          id\n          name\n          description\n        }\n      }\n      triageId\n      priorityId\n    }\n  }\n": types.GetMessagesDocument,
-    "\n  query GetMessageForTriage($messageId: uuid!) {\n    messagesByPk(id: $messageId) {\n      id\n      content\n      sender\n      receiver\n      senderDetail\n      receiverDetail\n      medium: mediumId\n      time\n      divisions {\n        division {\n          id\n          name\n          description\n        }\n      }\n      createdAt\n      updatedAt\n      deletedAt\n      triageId\n      priorityId\n      journal {\n        incident {\n          divisions {\n            id\n            name\n            description\n          }\n        }\n      }\n    }\n  }\n": types.GetMessageForTriageDocument,
-    "\n    mutation InsertMessage(\n      $journalId: uuid\n      $sender: String\n      $receiver: String\n      $time: timestamptz\n      $content: String\n      $receiverDetail: String\n      $senderDetail: String\n      $medium: MediumEnum\n    ) {\n      insertMessagesOne(\n        object: {\n          content: $content\n          journalId: $journalId\n          receiver: $receiver\n          sender: $sender\n          time: $time\n          mediumId: $medium\n          senderDetail: $senderDetail\n          receiverDetail: $receiverDetail\n        }\n      ) {\n        id\n        createdAt\n        content\n        receiver\n        sender\n        senderDetail\n        receiverDetail\n        medium: mediumId\n        time\n        updatedAt\n        triageId\n        priorityId\n        divisions {\n          division {\n            name\n          }\n        }\n        deletedAt\n      }\n    }\n  ": types.InsertMessageDocument,
-    "\n    mutation UpdateMessage(\n      $messageId: uuid!\n      $content: String\n      $sender: String\n      $receiver: String\n      $time: timestamptz\n      $receiverDetail: String\n      $senderDetail: String\n      $medium: MediumEnum\n    ) {\n      updateMessagesByPk(\n        pkColumns: { id: $messageId }\n        _set: {\n          content: $content\n          sender: $sender\n          receiver: $receiver\n          time: $time\n          mediumId: $medium\n          senderDetail: $senderDetail\n          receiverDetail: $receiverDetail\n        }\n      ) {\n        id\n        createdAt\n        content\n        receiver\n        sender\n        senderDetail\n        receiverDetail\n        medium: mediumId\n        time\n        updatedAt\n        triageId\n        priorityId\n        divisions {\n          division {\n            name\n          }\n        }\n        deletedAt\n      }\n    }\n  ": types.UpdateMessageDocument,
-    "\n  mutation SaveMessageTriage(\n    $messageId: uuid!\n    $priority: PriorityStatusEnum\n    $triage: TriageStatusEnum\n    $messageDivisions: [MessageDivisionInsertInput!]!\n  ) {\n    deleteMessageDivision(where: { messageId: { _eq: $messageId } }) {\n      affectedRows\n    }\n    insertMessageDivision(objects: $messageDivisions) {\n      affectedRows\n    }\n    updateMessagesByPk(\n      pkColumns: { id: $messageId }\n      _set: { priorityId: $priority, triageId: $triage }\n    ) {\n      id\n      divisions {\n        division {\n          name\n        }\n      }\n      triageId\n      priorityId\n    }\n  }\n": types.SaveMessageTriageDocument,
-};
-
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- *
- *
- * @example
- * ```ts
- * const query = graphql(`query GetUser($id: ID!) { user(id: $id) { name } }`);
- * ```
- *
- * The query argument is unknown!
- * Please regenerate the types.
- */
-export function graphql(source: string): unknown;
-
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n    query FetchIncidents {\n      incidents(orderBy: { createdAt: DESC }, where: { deletedAt: { _isNull: true } }) {\n        id\n        name\n        createdAt\n        updatedAt\n        deletedAt\n        closedAt\n        location {\n          name\n          coordinates\n        }\n      }\n    }\n  "): (typeof documents)["\n    query FetchIncidents {\n      incidents(orderBy: { createdAt: DESC }, where: { deletedAt: { _isNull: true } }) {\n        id\n        name\n        createdAt\n        updatedAt\n        deletedAt\n        closedAt\n        location {\n          name\n          coordinates\n        }\n      }\n    }\n  "];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetIncidentDetail($incidentId: uuid!) {\n    incidentsByPk(id: $incidentId) {\n      id\n      name\n      createdAt\n      closedAt\n      updatedAt\n      location {\n        id\n        name\n        coordinates\n      }\n      divisions {\n        id\n        name\n        description\n      }\n      journals {\n        id\n        name\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetIncidentDetail($incidentId: uuid!) {\n    incidentsByPk(id: $incidentId) {\n      id\n      name\n      createdAt\n      closedAt\n      updatedAt\n      location {\n        id\n        name\n        coordinates\n      }\n      divisions {\n        id\n        name\n        description\n      }\n      journals {\n        id\n        name\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation InsertIncident(\n    $name: String!\n    $location: String\n    $divisions: [DivisionsInsertInput!]!\n    $journalName: String\n    $layerName: String\n  ) {\n    insertIncidentsOne(\n      object: {\n        name: $name\n        location: { data: { name: $location } }\n        journals: { data: { name: $journalName } }\n        layers: { data: { name: $layerName } }\n        divisions: { data: $divisions }\n      }\n    ) {\n      id\n      name\n      journals {\n        id\n        name\n      }\n      divisions {\n        name\n        id\n        description\n      }\n      layers {\n        name\n        id\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation InsertIncident(\n    $name: String!\n    $location: String\n    $divisions: [DivisionsInsertInput!]!\n    $journalName: String\n    $layerName: String\n  ) {\n    insertIncidentsOne(\n      object: {\n        name: $name\n        location: { data: { name: $location } }\n        journals: { data: { name: $journalName } }\n        layers: { data: { name: $layerName } }\n        divisions: { data: $divisions }\n      }\n    ) {\n      id\n      name\n      journals {\n        id\n        name\n      }\n      divisions {\n        name\n        id\n        description\n      }\n      layers {\n        name\n        id\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation UpdateIncident(\n    $incidentId: uuid!\n    $name: String!\n    $location: String!\n    $locationId: uuid!\n    $divisions: [DivisionsInsertInput!]!\n  ) {\n    updateLocationsByPk(pkColumns: { id: $locationId }, _set: { name: $location }) {\n      id\n      name\n    }\n    insertDivisions(\n      objects: $divisions\n      onConflict: { constraint: divisions_name_incident_id_key, updateColumns: [description, name] }\n    ) {\n      affectedRows\n    }\n    updateIncidentsByPk(pkColumns: { id: $incidentId }, _set: { name: $name }) {\n      id\n      name\n      journals {\n        id\n        name\n      }\n      divisions {\n        name\n        id\n        description\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation UpdateIncident(\n    $incidentId: uuid!\n    $name: String!\n    $location: String!\n    $locationId: uuid!\n    $divisions: [DivisionsInsertInput!]!\n  ) {\n    updateLocationsByPk(pkColumns: { id: $locationId }, _set: { name: $location }) {\n      id\n      name\n    }\n    insertDivisions(\n      objects: $divisions\n      onConflict: { constraint: divisions_name_incident_id_key, updateColumns: [description, name] }\n    ) {\n      affectedRows\n    }\n    updateIncidentsByPk(pkColumns: { id: $incidentId }, _set: { name: $name }) {\n      id\n      name\n      journals {\n        id\n        name\n      }\n      divisions {\n        name\n        id\n        description\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation CloseIncident($incidentId: uuid, $closedAt: timestamptz) {\n    updateJournals(\n      where: { incident: { id: { _eq: $incidentId } }, closedAt: { _isNull: true } }\n      _set: { closedAt: $closedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        closedAt\n      }\n    }\n    updateIncidents(where: { id: { _eq: $incidentId } }, _set: { closedAt: $closedAt }) {\n      affectedRows\n      returning {\n        id\n        closedAt\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation CloseIncident($incidentId: uuid, $closedAt: timestamptz) {\n    updateJournals(\n      where: { incident: { id: { _eq: $incidentId } }, closedAt: { _isNull: true } }\n      _set: { closedAt: $closedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        closedAt\n      }\n    }\n    updateIncidents(where: { id: { _eq: $incidentId } }, _set: { closedAt: $closedAt }) {\n      affectedRows\n      returning {\n        id\n        closedAt\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation DeleteIncident($incidentId: uuid, $deletedAt: timestamptz) {\n    updateJournals(\n      where: { incident: { id: { _eq: $incidentId } }, deletedAt: { _isNull: true } }\n      _set: { deletedAt: $deletedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        deletedAt\n      }\n    }\n    updateIncidents(\n      where: {\n        id: { _eq: $incidentId }\n        deletedAt: { _isNull: true }\n        closedAt: { _isNull: false }\n      }\n      _set: { deletedAt: $deletedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        deletedAt\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation DeleteIncident($incidentId: uuid, $deletedAt: timestamptz) {\n    updateJournals(\n      where: { incident: { id: { _eq: $incidentId } }, deletedAt: { _isNull: true } }\n      _set: { deletedAt: $deletedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        deletedAt\n      }\n    }\n    updateIncidents(\n      where: {\n        id: { _eq: $incidentId }\n        deletedAt: { _isNull: true }\n        closedAt: { _isNull: false }\n      }\n      _set: { deletedAt: $deletedAt }\n    ) {\n      affectedRows\n      returning {\n        id\n        deletedAt\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetJournals($incidentId: uuid) {\n    incidents(where: { id: { _eq: $incidentId } }) {\n      id\n      name\n      journals(orderBy: { createdAt: ASC }) {\n        id\n        name\n        createdAt\n        updatedAt\n        closedAt\n        deletedAt\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetJournals($incidentId: uuid) {\n    incidents(where: { id: { _eq: $incidentId } }) {\n      id\n      name\n      journals(orderBy: { createdAt: ASC }) {\n        id\n        name\n        createdAt\n        updatedAt\n        closedAt\n        deletedAt\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation InsertJournal($name: String!, $incidentId: uuid!) {\n    insertJournalsOne(object: { incidentId: $incidentId, name: $name }) {\n      id\n      name\n      createdAt\n      updatedAt\n      closedAt\n      deletedAt\n    }\n  }\n"): (typeof documents)["\n  mutation InsertJournal($name: String!, $incidentId: uuid!) {\n    insertJournalsOne(object: { incidentId: $incidentId, name: $name }) {\n      id\n      name\n      createdAt\n      updatedAt\n      closedAt\n      deletedAt\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n    mutation CloseJournal($journalId: uuid, $closedAt: timestamptz) {\n      updateJournals(where: { id: { _eq: $journalId } }, _set: { closedAt: $closedAt }) {\n        affectedRows\n        returning {\n          id\n          closedAt\n        }\n      }\n    }\n  "): (typeof documents)["\n    mutation CloseJournal($journalId: uuid, $closedAt: timestamptz) {\n      updateJournals(where: { id: { _eq: $journalId } }, _set: { closedAt: $closedAt }) {\n        affectedRows\n        returning {\n          id\n          closedAt\n        }\n      }\n    }\n  "];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetLayers($incidentId: uuid!) {\n    layers(where: { incidentId: { _eq: $incidentId } }) {\n      id\n      name\n      features {\n        id\n        geometry\n        properties\n        createdAt\n        updatedAt\n        deletedAt\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetLayers($incidentId: uuid!) {\n    layers(where: { incidentId: { _eq: $incidentId } }) {\n      id\n      name\n      features {\n        id\n        geometry\n        properties\n        createdAt\n        updatedAt\n        deletedAt\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation AddFeature($layerId: uuid!, $id: uuid!, $geometry: jsonb, $properties: jsonb) {\n    insertFeaturesOne(\n      object: { layerId: $layerId, id: $id, geometry: $geometry, properties: $properties }\n    ) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"): (typeof documents)["\n  mutation AddFeature($layerId: uuid!, $id: uuid!, $geometry: jsonb, $properties: jsonb) {\n    insertFeaturesOne(\n      object: { layerId: $layerId, id: $id, geometry: $geometry, properties: $properties }\n    ) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation UpdateFeature($id: uuid!, $geometry: jsonb, $properties: jsonb) {\n    updateFeaturesByPk(\n      pkColumns: { id: $id }\n      _set: { geometry: $geometry, properties: $properties }\n    ) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"): (typeof documents)["\n  mutation UpdateFeature($id: uuid!, $geometry: jsonb, $properties: jsonb) {\n    updateFeaturesByPk(\n      pkColumns: { id: $id }\n      _set: { geometry: $geometry, properties: $properties }\n    ) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation DeleteFeature($id: uuid!, $deletedAt: timestamptz) {\n    updateFeaturesByPk(pkColumns: { id: $id }, _set: { deletedAt: $deletedAt }) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"): (typeof documents)["\n  mutation DeleteFeature($id: uuid!, $deletedAt: timestamptz) {\n    updateFeaturesByPk(pkColumns: { id: $id }, _set: { deletedAt: $deletedAt }) {\n      id\n      geometry\n      properties\n      createdAt\n      updatedAt\n      deletedAt\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation AddLayer($incidentId: uuid!, $name: String!) {\n    insertLayersOne(object: { incidentId: $incidentId, name: $name }) {\n      id\n    }\n  }\n"): (typeof documents)["\n  mutation AddLayer($incidentId: uuid!, $name: String!) {\n    insertLayersOne(object: { incidentId: $incidentId, name: $name }) {\n      id\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetMessages($journalId: uuid!) {\n    journalsByPk(id: $journalId) {\n      incident {\n        id\n        divisions {\n          id\n          name\n          description\n        }\n      }\n    }\n    messages(\n      where: { journal: { id: { _eq: $journalId } }, deletedAt: { _isNull: true } }\n      orderBy: { time: DESC }\n    ) {\n      id\n      content\n      sender\n      receiver\n      senderDetail\n      receiverDetail\n      medium: mediumId\n      time\n      createdAt\n      updatedAt\n      deletedAt\n      divisions {\n        division {\n          id\n          name\n          description\n        }\n      }\n      triageId\n      priorityId\n    }\n  }\n"): (typeof documents)["\n  query GetMessages($journalId: uuid!) {\n    journalsByPk(id: $journalId) {\n      incident {\n        id\n        divisions {\n          id\n          name\n          description\n        }\n      }\n    }\n    messages(\n      where: { journal: { id: { _eq: $journalId } }, deletedAt: { _isNull: true } }\n      orderBy: { time: DESC }\n    ) {\n      id\n      content\n      sender\n      receiver\n      senderDetail\n      receiverDetail\n      medium: mediumId\n      time\n      createdAt\n      updatedAt\n      deletedAt\n      divisions {\n        division {\n          id\n          name\n          description\n        }\n      }\n      triageId\n      priorityId\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  query GetMessageForTriage($messageId: uuid!) {\n    messagesByPk(id: $messageId) {\n      id\n      content\n      sender\n      receiver\n      senderDetail\n      receiverDetail\n      medium: mediumId\n      time\n      divisions {\n        division {\n          id\n          name\n          description\n        }\n      }\n      createdAt\n      updatedAt\n      deletedAt\n      triageId\n      priorityId\n      journal {\n        incident {\n          divisions {\n            id\n            name\n            description\n          }\n        }\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetMessageForTriage($messageId: uuid!) {\n    messagesByPk(id: $messageId) {\n      id\n      content\n      sender\n      receiver\n      senderDetail\n      receiverDetail\n      medium: mediumId\n      time\n      divisions {\n        division {\n          id\n          name\n          description\n        }\n      }\n      createdAt\n      updatedAt\n      deletedAt\n      triageId\n      priorityId\n      journal {\n        incident {\n          divisions {\n            id\n            name\n            description\n          }\n        }\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n    mutation InsertMessage(\n      $journalId: uuid\n      $sender: String\n      $receiver: String\n      $time: timestamptz\n      $content: String\n      $receiverDetail: String\n      $senderDetail: String\n      $medium: MediumEnum\n    ) {\n      insertMessagesOne(\n        object: {\n          content: $content\n          journalId: $journalId\n          receiver: $receiver\n          sender: $sender\n          time: $time\n          mediumId: $medium\n          senderDetail: $senderDetail\n          receiverDetail: $receiverDetail\n        }\n      ) {\n        id\n        createdAt\n        content\n        receiver\n        sender\n        senderDetail\n        receiverDetail\n        medium: mediumId\n        time\n        updatedAt\n        triageId\n        priorityId\n        divisions {\n          division {\n            name\n          }\n        }\n        deletedAt\n      }\n    }\n  "): (typeof documents)["\n    mutation InsertMessage(\n      $journalId: uuid\n      $sender: String\n      $receiver: String\n      $time: timestamptz\n      $content: String\n      $receiverDetail: String\n      $senderDetail: String\n      $medium: MediumEnum\n    ) {\n      insertMessagesOne(\n        object: {\n          content: $content\n          journalId: $journalId\n          receiver: $receiver\n          sender: $sender\n          time: $time\n          mediumId: $medium\n          senderDetail: $senderDetail\n          receiverDetail: $receiverDetail\n        }\n      ) {\n        id\n        createdAt\n        content\n        receiver\n        sender\n        senderDetail\n        receiverDetail\n        medium: mediumId\n        time\n        updatedAt\n        triageId\n        priorityId\n        divisions {\n          division {\n            name\n          }\n        }\n        deletedAt\n      }\n    }\n  "];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n    mutation UpdateMessage(\n      $messageId: uuid!\n      $content: String\n      $sender: String\n      $receiver: String\n      $time: timestamptz\n      $receiverDetail: String\n      $senderDetail: String\n      $medium: MediumEnum\n    ) {\n      updateMessagesByPk(\n        pkColumns: { id: $messageId }\n        _set: {\n          content: $content\n          sender: $sender\n          receiver: $receiver\n          time: $time\n          mediumId: $medium\n          senderDetail: $senderDetail\n          receiverDetail: $receiverDetail\n        }\n      ) {\n        id\n        createdAt\n        content\n        receiver\n        sender\n        senderDetail\n        receiverDetail\n        medium: mediumId\n        time\n        updatedAt\n        triageId\n        priorityId\n        divisions {\n          division {\n            name\n          }\n        }\n        deletedAt\n      }\n    }\n  "): (typeof documents)["\n    mutation UpdateMessage(\n      $messageId: uuid!\n      $content: String\n      $sender: String\n      $receiver: String\n      $time: timestamptz\n      $receiverDetail: String\n      $senderDetail: String\n      $medium: MediumEnum\n    ) {\n      updateMessagesByPk(\n        pkColumns: { id: $messageId }\n        _set: {\n          content: $content\n          sender: $sender\n          receiver: $receiver\n          time: $time\n          mediumId: $medium\n          senderDetail: $senderDetail\n          receiverDetail: $receiverDetail\n        }\n      ) {\n        id\n        createdAt\n        content\n        receiver\n        sender\n        senderDetail\n        receiverDetail\n        medium: mediumId\n        time\n        updatedAt\n        triageId\n        priorityId\n        divisions {\n          division {\n            name\n          }\n        }\n        deletedAt\n      }\n    }\n  "];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation SaveMessageTriage(\n    $messageId: uuid!\n    $priority: PriorityStatusEnum\n    $triage: TriageStatusEnum\n    $messageDivisions: [MessageDivisionInsertInput!]!\n  ) {\n    deleteMessageDivision(where: { messageId: { _eq: $messageId } }) {\n      affectedRows\n    }\n    insertMessageDivision(objects: $messageDivisions) {\n      affectedRows\n    }\n    updateMessagesByPk(\n      pkColumns: { id: $messageId }\n      _set: { priorityId: $priority, triageId: $triage }\n    ) {\n      id\n      divisions {\n        division {\n          name\n        }\n      }\n      triageId\n      priorityId\n    }\n  }\n"): (typeof documents)["\n  mutation SaveMessageTriage(\n    $messageId: uuid!\n    $priority: PriorityStatusEnum\n    $triage: TriageStatusEnum\n    $messageDivisions: [MessageDivisionInsertInput!]!\n  ) {\n    deleteMessageDivision(where: { messageId: { _eq: $messageId } }) {\n      affectedRows\n    }\n    insertMessageDivision(objects: $messageDivisions) {\n      affectedRows\n    }\n    updateMessagesByPk(\n      pkColumns: { id: $messageId }\n      _set: { priorityId: $priority, triageId: $triage }\n    ) {\n      id\n      divisions {\n        division {\n          name\n        }\n      }\n      triageId\n      priorityId\n    }\n  }\n"];
-
-export function graphql(source: string) {
-  return (documents as any)[source] ?? {};
-}
-
-export type DocumentType<TDocumentNode extends DocumentNode<any, any>> = TDocumentNode extends DocumentNode<  infer TType,  any>  ? TType  : never;
