--- ui/src/gql/index.ts
@@ -1,1 +1,0 @@
-export * from "./gql";
