+++ ui/src/types/context.ts
@@ -1,5 +1,4 @@
 import type { Incident } from "./incident";
-import type { Journal } from "./journal";
 export interface UserState {
   isLoggedin: boolean;
   username: string;
@@ -8,5 +7,4 @@ export interface UserState {
 
 export interface IncidentContext {
   incident: Incident | null;
-  journal: Journal | null;
 }
