+++ ui/src/types/index.ts
@@ -1,4 +1,4 @@
 export type { IncidentContext, UserState } from "./context";
 export type { Incident, Location } from "./incident";
-export type { Division, DivisionList, Journal, Message, Priority, Triage } from "./journal";
+export type { Division, DivisionList, Message, Priority, Triage } from "./journal";
 export { Medium, PriorityStatus, TriageStatus } from "./journal";
