+++ ui/src/types/journal.ts
@@ -1,5 +1,3 @@
-import type { Incident } from "./incident";
-
 export enum TriageStatus {
   Pending = "PENDING",
   Triaged = "DONE",
@@ -57,12 +55,3 @@ export interface Division {
   description: string;
 }
 
-export interface Journal {
-  id: string;
-  name: string;
-  incident: Incident;
-  createdAt: Date;
-  updatedAt: Date;
-  closedAt: Date;
-  deletedAt: Date;
-}
