+++ ui/src/utils/IncidentContext.tsx
@@ -7,17 +7,14 @@ import {
   useReducer,
 } from "react";
 import { useParams } from "react-router";
-import type { Incident, IncidentContext as IncidentContextState, Journal } from "types";
+import type { Incident, IncidentContext as IncidentContextState } from "types";
 import { useIncidentDetails } from "api";
 
 const initialState: IncidentContextState = {
   incident: null,
-  journal: null,
 };
 
-type IncidentAction =
-  | { type: "SET_INCIDENT"; payload: Incident | null }
-  | { type: "SET_JOURNAL"; payload: Journal | null };
+type IncidentAction = { type: "SET_INCIDENT"; payload: Incident | null };
 
 const incidentReducer = (
   state: IncidentContextState,
@@ -26,8 +23,6 @@ const incidentReducer = (
   switch (action.type) {
     case "SET_INCIDENT":
       return { ...state, incident: action.payload };
-    case "SET_JOURNAL":
-      return { ...state, journal: action.payload };
     default:
       return state;
   }
@@ -49,7 +44,7 @@ const IncidentContextProvider = ({ children }: { children: ReactNode }) => {
 };
 
 const IncidentContextSetter = () => {
-  const { incidentId, journalId } = useParams();
+  const { incidentId } = useParams();
   const { state, dispatch } = useContext(IncidentContext);
   const result = useIncidentDetails(incidentId);
 
@@ -57,7 +52,6 @@ const IncidentContextSetter = () => {
     if (!incidentId) {
       if (state.incident !== null) {
         dispatch({ type: "SET_INCIDENT", payload: null });
-        dispatch({ type: "SET_JOURNAL", payload: null });
       }
       return;
     }
@@ -68,14 +62,7 @@ const IncidentContextSetter = () => {
     if (state.incident?.id !== incident.id) {
       dispatch({ type: "SET_INCIDENT", payload: incident });
     }
-
-    if (journalId) {
-      const journal = incident.journals.find((j) => j.id === journalId);
-      if (journal && state.journal?.id !== journal.id) {
-        dispatch({ type: "SET_JOURNAL", payload: journal });
-      }
-    }
-  }, [incidentId, journalId, result.status, result.data, state.incident, state.journal, dispatch]);
+  }, [incidentId, result.status, result.data, state.incident, dispatch]);
 
   return null;
 };
